迁移福州

+++ HouseCrawler/AirflowAdmin/dags/NewHouseZQ.py
@@ -126,7 +126,7 @@ for item in cur:
                                       'BuildingUUID': str(item['BuildingUUID'])}}
             building_info_list.append(building_info)
 
-index_skip = int(math.ceil(len(building_info_list) / float(3))) + 1
+index_skip = int(math.ceil(len(building_info_list) / float(4))) + 1
 for cur, index in enumerate(list(range(0, len(building_info_list), index_skip))):
     t3 = PythonOperator(
         task_id='LoadBuildingInfoZQ_%s' % cur,
